// examples/workflow/main/src/lib/model/enum/actions/update-invoice-header-action-editor-params.enum.ts
/*
 * Copyright(c) RIB Software GmbH
 */

/**
 * Input/Output parameters for the action 'Update Invoice Header Action'
 */
export enum UpdateInvoiceHeaderActionParamsEnum {
	Code = 'Code',
	InvoiceId = 'InvoiceId',
	NewCode = 'NewCode',
	Remark = 'Remark',
	Userdefined1 = 'Userdefined1',
	Userdefined2 = 'Userdefined2',
	Userdefined3 = 'Userdefined3',
	Userdefined4 = 'Userdefined4',
	Userdefined5 = 'Userdefined5',
	Result = 'Result',
}